+++ src/components/utilities/Defense.vue
@@ -3,7 +3,7 @@
     <LoadingSpinner v-if="loading" />
 
     <h2 class="accordion-header" :id="id + '-heading'">
-      <button class="accordion-button" type="button" @click="defense.userData.isCollapsed = !defense.userData?.isCollapsed" :class="{ collapsed }" data-bs-toggle="collapse" :data-bs-target="'#' + id" :aria-expanded="!collapsed" :aria-controls="id">
+      <button class="accordion-button" type="button" @click="defense.userData.isCollapsed = setupDefenses ? defense.userData?.isCollapsed : !defense.userData?.isCollapsed" :class="{ collapsed }" data-bs-toggle="collapse" :data-bs-target="'#' + id" :aria-expanded="!collapsed" :aria-controls="id">
         <span class="d-flex justify-content-between w-100">
           <span class="defense-label">{{ defense.userData?.label }}</span>
 
@@ -48,6 +48,7 @@
                 <span v-if="isDev" class="w-100 defense-info__header-stats__stat"><strong>Attack damage:</strong> {{ Math.round(attackDamage).toLocaleString('en-US') }}</span>
                 <span class="w-100 defense-info__header-stats__stat"><strong>Defense Power:</strong> {{ Math.round(defensePower) }}</span>
                 <span class="w-100 defense-info__header-stats__stat"><strong>Defense Health:</strong> {{ Math.round(defenseHealth) }}</span>
+                <span class="w-100 defense-info__header-stats__stat"><strong>Defense Rate:</strong> {{ attackRate }} ({{ attackRatePercentage }}%)</span>
                 <span class="w-100 defense-info__header-stats__stat"><strong>Crit chance:</strong> {{ (criticalChance * 100).toFixed(2) }}%</span>
                 <span class="w-100 defense-info__header-stats__stat"><strong>Crit damage:</strong> {{ (criticalDamage * 100).toFixed(2) }}%</span>
               </div>
@@ -72,7 +73,7 @@
             <div class="defense-info__pet">
               <div class="row">
                 <div class="col-md-6">
-                  <DefenseRelic v-model="defense.userData.relic" :defenseCompatibility="defense.defenseData?.id" :hide-mods="true" />
+                  <DefenseRelic v-model="defense.userData.relic" :defenseCompatibility="defense.userData.id" :hide-mods="true" />
                 </div>
                 <div class="col-md-6">
                   <Pet v-model="defense.userData.pet" />
@@ -83,7 +84,7 @@
             <hr />
 
             <div class="defense-info__relic">
-              <DefenseRelic v-model="defense.userData.relic" :defenseCompatibility="defense.defenseData?.id" :hide-relic="true" />
+              <DefenseRelic v-model="defense.userData.relic" :defenseCompatibility="defense.userData.id" :hide-relic="true" />
 
               <i v-if="hasDiverseMods">For proper testing diverse mods, use defense setups (see section below)</i>
             </div>
@@ -91,7 +92,7 @@
             <hr />
 
             <div class="defense-info__shards">
-              <Shards v-model="defense.userData.shards" :defenseCompatibility="defense.defenseData?.id" />
+              <Shards v-model="defense.userData.shards" :defenseCompatibility="defense.userData.id" />
             </div>
 
             <hr />
@@ -107,7 +108,7 @@
 </template>
 
 <script setup lang="ts">
-import { ref, watch, defineProps, defineEmits, onMounted } from "vue";
+import { ref, watch, defineProps, defineEmits, onMounted, computed } from "vue";
 import type { PropType } from "vue";
 import type {
   DefenseRootInterface,
@@ -149,7 +150,7 @@ const { debounce } = useDebounce()
 const { loading } = storeToRefs(googleSpreadsheetDataStore)
 const { deleteDefense } = userStore
 const { ancientPowerPoints, isDev } = storeToRefs(userStore);
-const { totalDps, tooltipDps, attackDamage, defensePower, defenseHealth, criticalDamage, criticalChance, calculateDefensePower, isBuffDefense } = useDefenseCalculations()
+const { totalDps, tooltipDps, attackDamage, attackRate, defensePower, defenseHealth, criticalDamage, criticalChance, calculateDefensePower, isBuffDefense } = useDefenseCalculations()
 const { getModById } = useModStore()
 const { getShardById } = useShardStore()
 
@@ -173,6 +174,7 @@ const defenseLevel = ref<number>(1)
 const hasDiverseMods = ref<boolean>(false)
 const userDefenseMods = ref<DefenseModData[]>([])
 const userDefenseShards = ref<DefenseShardData[]>([])
+const attackRatePercentage = computed(() => Math.round((attackRate.value - (defense.defenseData?.baseAttackRate ?? 0)) / ((defense.defenseData?.maxAttackRate ?? 0) - (defense.defenseData?.baseAttackRate ?? 0)) * 100))
 
 function recalculate(): void {
   if (!defense.userData) return
@@ -236,7 +238,7 @@ onMounted((): void => {
   id.value = 'id' + Math.floor((1 + Math.random()) * 0x10000)
       .toString(16)
       .substring(1)
-      .toLowerCase();
+      .toLowerCase() + defense.incrementId
 
   // Await the loading of defenseData before initializing calculations
   const interval: any = setInterval((): void => {