+++ src/stores/UserData.ts
@@ -17,6 +17,7 @@ export const useUserDataStore = defineStore('userDataStore', () => {
     const dataMigration = new DataMigrations
 
     const colorMode = ref<string>(localStorage.getItem('colorMode') ?? 'light')
+    const lastVisitedVersion = ref<string>(localStorage.getItem('lastVisitedVersion') ?? 'v0.0.0')
     const isDev = ref<boolean>(localStorage.getItem('isDev') === 'true')
     const defenses = ref<UserDataStoreDefenseInterface[]>(getDefenses())
     const defenseSetups = ref<UserDefenseSetupInterface[]>(getDefenseSetups())
@@ -145,12 +146,15 @@ export const useUserDataStore = defineStore('userDataStore', () => {
         localStorage.setItem('ancientResetPoints', JSON.stringify(ancientPowerPoints.value))
     }, { deep: true })
 
-    // Persist ancient power data on change
     watch(colorMode, () => {
         localStorage.setItem('colorMode', colorMode.value)
     })
 
-    return { isDev, colorMode, defenses, defenseSetups, ancientPowerPoints, deleteDefense, deleteDefenseSetup, getNextDefenseIncrementId, getNextDefenseSetupIncrementId, importDefenses, importDefenseSetups }
+    watch(lastVisitedVersion, () => {
+        localStorage.setItem('lastVisitedVersion', lastVisitedVersion.value)
+    })
+
+    return { isDev, colorMode, lastVisitedVersion, defenses, defenseSetups, ancientPowerPoints, deleteDefense, deleteDefenseSetup, getNextDefenseIncrementId, getNextDefenseSetupIncrementId, importDefenses, importDefenseSetups }
 })
 
 export function getDefaultSetupModifiers(): DefenseSetupModifiersInterface {